added components in home.tsx
